+++ config/initializers/load_libraries.rb
@@ -13,7 +13,6 @@ require 'typhoeus'
 require 'post_presenter'
 
 # Our libs
-require 'collect_user_photos'
 require 'diaspora'
 require 'direction_detector'
 require 'email_inviter'